requirements.txt
Django==4.0.3
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
ipython==8.1.1
mysqlclient==2.1.0
